Gulp build-upload task now uses TeamCity SSH agent
The TeamCity build configuration now runs an SSH agent that will provide
the appropriate SSH key for the rsync, so we no longer specify it with
the -i command. We also no longer use `sudo rsync` on the remote end, as
we have changed the setup for uploading to use a dedicated upload_areas
directory with symlinks to the right locations.